Metodos para obtener series por temporadas por artista o por titulo

+++ src/main/java/ar/com/ada/api/pootflix/model/request/PeliculaRequest.java
@@ -7,7 +7,7 @@ public class PeliculaRequest {
     public String premio;
     public Integer genero;
     public ObjectId idActor;
-    public Integer calificacion;
+    public Double calificacion;
     public boolean filmadaenImax;
     public Integer paisOrigen;
     public Integer anio;
